Rename helpers to mocks
